SGTools: Optimizations for requestRelease.
Remove locking for requestRelease, as it is only called when the store
is being cleared.  To prevent misuse, change it to private and allow
a call from DataStore.  requestRelease also does not do a release itself;
rather, the caller is expected to do so if it returns true.  This is to 
prepare for a reorganization of DataStore.


Former-commit-id: 5e3b8294d86742146823523d3aa9bec5b5ec7a95