Avoid assuming a beamspot AOD merge step
The T0 beamspot jobs are being reconfigured to skip the AOD merge step and run
the beamspotproc step directly on the transient AODs. This means that the
dataset names will change slightly according to this example:

   OLD: data17_13TeV.<runnr>.calibration_BeamSpot.merge.AOD.c1150_m1879_c1113.beamspotproc.task
   NEW: data17_13TeV.<runnr>.calibration_BeamSpot.recon.AOD.c1150_c1113.beamspotproc.task

This change captures the two remaining places where merge.AOD was hard-coded
and instead uses merge with an m tag is present, and recon otherwise.


Former-commit-id: 9d3ac54657d1f21cef963371dafec0b31d3ae986