Final edits to get unit test log comparisons working in SUSYTools
Removing the environment dump in some unit tests.  Extracting the input
files for the tests from the environment instead, so that we can
immediately see if the input file changed.  Adding exclusions for all
the other environmental variables that might be set seemed far too
excessive (and rather tricky to do for every user...)

Updating veto lists in CMakeLists file to make sure that the log
comparisons are fully working.  Added comments to explain the veto lines
that are now there.

Reduced to DEBUG a couple of outputs in the MET systematics tool, where
it literally prints out two environmental variables.  It prints out the
file that it finds, in the end, so I don't think this is really a useful
'INFO' level printout.  Will tag the METUtilities developers just in
case.


Former-commit-id: b6c98b7a44bd2ef7a990e457961b031ce33abbca