remove tracks associated to jets to reduce size
no jira

Tracks near jets are no longer marked to be saved by default to save disk space. Only STDM derivation that still keeps these jets is STDM9, which is used for jet substructure analyses which rely on these tracks. For all other derivations there should not be a problem.


Former-commit-id: 4f94e07718010dcb64682616b1585302428287b3