Adding one more veto
WP checker in SUSYTools makes a local config file that is then used by
the job.  This shows up in the run directory, whatever that is, which
changes from local to build to CI jobs.  Don't pay attention to log
lines that have that file on them, as their differences are red
herrings.


Former-commit-id: 79bb6cbf390700e0edd1401baa0b6b94fa910bc6