fix overlow within quadraticSumBW
Details:
convert all internal variables of this function to unsigned long int.
This ensures that v^2 values, where v is the input int, will fit
in the available bits.
Note that the firmware version is `UNSIGNED --> UNSIGNED`.

In software we have our internal manipulation of bits (for example see
L1TopoEvent/MetTOB.h that provides the inputs to this function), and
the function is `int --> int`. We can reasonably assume that the input
value is positive, although we might want to consider including a
check (see TODO).

Also changed the unit test input values to go from 0 to 32767 since
there are 16 bits available; 65535 was overflowing (int vs. uint)


Former-commit-id: f5791ff180bfbe0bfa6440eca9a2d4e1f791a332