AthenaBaseComps: MT message output level fixes.
The previous thread-safety fix to AthMessaging meant that we were retrieving
from a boost thread specific pointer every time we needed to do something
with the MsgStream --- including checking the message level in the ATH_MSG_
macros.  This was causing performance problems in reconstruction.
In addition, replicating the MsgStreams means that attempts
to set a non-default output level won't work properly in MT jobs
(will be set only for one thread).

Changed to cache the message level in AthMessaging, as an atomic that is
visible to all threads.  This can now be read by the ATH_MSG_ macros.
In addition, we use this to set the output level for the per-thread
MsgStream instrances.



Former-commit-id: e9b1a2381c3d95fd878b9ed8908f236f1ae7300e