Fixed the conversion from IOVTime to EventIDBase (ATEAM-432)
The fix is required for proper handling of time-indexed COOL folders with
infinite IOVs. Such IOV ranges will be converted into EventIDRange with the
upper time-stamp limit equal to std::numeric_limits<unsigned int>::max()-1.

This change is not sufficient for fixing the issues reported in ATEAM-432.
More fixes are expected in separate MR-s in the future. For example:
we also need to fix Digitization configuration, such that it will enable
the MT condition access infrastructure in digitization jobs.


Former-commit-id: a923481c660773293952c2894fc7e7888edc8793