One last tweak to paths for SUSYTools UTs
The paths for the nightly builds is slightly different in structure from
the paths for the CI builds, which means the veto strings used in the
log files need to be slightly different.

As pointed out in earlier MR, happy to adopt a central solution to this
issue if one exists.


Former-commit-id: 618c6891a671cd58abf49fe23a5c0ef7bc24623a