Add new HepMcParticleLink_p2 persistent format
This commit adds the new `HepMcParticleLink_p2` persistent format
that will in future be written out by jobs run `21.3` and `master`.
These changes are for forward compatibility only, so while a P->T
method is defined, the T->P method of `HepMcParticleLinkCnv_p2`
just throws an exception saying that writing `HepMcParticleLink_p2`
is not supported in this release. See ATLASSIM-3457.


Former-commit-id: 0cd6453e36785d5128d7d7b3bc50df525a372abb