Avoid building of an empty LVL1CTP::Lvl1Result in case no valid CTP result is available (ATR-17043)
In partially build events, like DataScouting events, no CTP ROBs may be available. In this case no
empty LVL1CTP::Lvl1Result should be added to StoreGate. If this record is found in StoreGate the
Trigger Decision Tool (TDT) will try to obtain from it the L1 trigger decision. If the record is invalid,
no L1 trigger decision will be available in analysis records, while in case the LVL1CTP::Lvl1Result is not
found in StoreGate the TDT will automatically retrieve the L1 decision from the event header and make the
approprite L1 decision is available in analysis records.


Former-commit-id: 224d79ed427e7a2480d6797bf0db04e8e718225d