Add support for new persistent formats using HepMcParticleLink_p2
This commit adds the new persistent formats for the `MuonDetSimEvent`
classes that will in future be written out by jobs run `21.3` and
`master`.  These changes are for forward compatibility only, so
while `persToTrans` methods are defined, the `transToPers` methods
just throw an exception saying that writing these persistent formats
is not supported in this release. See ATLASSIM-3457.


Former-commit-id: 89babcd24bdac5a58187275ef9b493af88279e38