Add support for new persistent formats using HepMcParticleLink_p2
This commit adds the new persistent formats for the `MuonDetSimEvent`
classes that will in future be written out by jobs run `21.3` and
`master`.  These changes are for forward compatibility only, so
while `persToTrans` methods are defined, the `transToPers` methods
just throw an exception saying that writing these persistent formats
is not supported in this release. See ATLASSIM-3457.


Former-commit-id: fbe02dc419d0ed59585c86a3aa8606b4ba4198d5