Add support for new persistent formats using HepMcParticleLink_p2
This commit adds the new persistent formats for the `MuonDetSimData`
and `CscSimData` (SDO) classes that will in future be written out by
jobs run `21.3` and `master`.  These changes are for forward
compatibility only, so while `persToTrans` method is defined, the
`transToPers` method just throws an exception saying that writing
this persistent format is not supported in this release.
See ATLASSIM-3457.


Former-commit-id: 529fc8839f9b15f441d62bbc8d60c0469e83f04f