Add support for new persistent formats using HepMcParticleLink_p2
This commit adds the new persistent formats for the `LUCID_SimHit`
class that will in future be written out by jobs run `21.3` and
`master`.  These changes are for forward compatibility only, so
while the `persToTrans` method is defined, the `transToPers` method
just throws an exception saying that writing this persistent format
is not supported in this release. See ATLASSIM-3457.


Former-commit-id: 89fe32c1f61af1fad2de35ae180fafd239289179