Missed definition of kernel
In the migration, completely missed that kernel was an undefined
variable after the changes had been made.  Making sure that we get it
from a sensible place and that it is defined.  The algorithm is very
fast, so there should be minimal overhead from adding it centrally (this
is also standard practice for other metadata tools).


Former-commit-id: d51d13ffed1891c549542bb6e73fd4dc90ceb536