Updating default treatment of geometry and conditions
This change only affects jobs that are running workflows that *do not*
set the geometry and conditions by hand (so not evgen or simulation) and
that cannot use the standard autoconfiguration (so not standard
digi/reco/derivations).  The cases that I've come across so far are
bytestream, where I believe we always set the defaults for running on
data, and derivation production from EVNT files, where there is no
useful geometry information available in the file to peek at.

The old code tested for the existence of a file on afs -- whether that
file was being used or not.  This caused differences between jobs
running locally at CERN, where afs is available, and on the grid, where
an afs connection to CERN is not always available.  The code that was
there appears to have been added to deal with a migration of test files
many years ago, and was only partly updated when old geometries were
being phased out in R21.  As a result, TRUTH derivation merge jobs were
failing on the grid.  This is described in ATLASRECTS-4191.

The update to the conditions tag, I don't believe will harm anything,
since the only work flows I can think of that are usuing this code and
are 'standard' workflows don't actually care about the conditions they
load.  At some point in the future it might be nice to try to get the
jobs that deal with TRUTH derivations to completely avoid dealing with
geometry and conditions all together, though this required quite a lot
of configuration tweaking in old releases before the xAODs came in.

This is to be discussed, of course, with a broader audience!


Former-commit-id: a3a691c15e01de99e2b61c3d0134836aa6b2eaa0