Adding TruthPVCollectionMaker and b/c-hadron options
Solving two more outstanding requests for truth tools / truth
derivations.  To the TruthDecayCollectionMaker, adding options for
keeping all b- and/or c-hadrons, along with all downstream particles.
This should provide a solution to the HF group, which wants exactly
these particles but currently is keeping the entire truth record.

The second is a new TruthPVCollectionMaker, which makes a small truth
primary vertex collection, with one vertex for each event in the event
collection.  This should allow people to find the location (in z) of the
true vertices of the in-time pileup events (depending on the
configuration of their MC sample).

In both cases, small python snippets are provided for configuring the
tools and adding them to a derivation.  Users can directly involk these
functions, or dream up their own version.


Former-commit-id: d92a1abedbe29c8a40e1cac2b48836ce2de68064