adjust MuonTrackingGeometry for NSW (ATLASSIM-3458)
At some point the volume naming in the input station.xml file
changed. The MuonTrackingGeometry is sensitive to this naming
and needs adjustment. Furthermore the shapes in MuonGeoModel
for some NSW volumes changed. Again the tracking geometry needs
to be changed to cope with that. Further code cleanup was
performed as well. Running q431 was running w/o crashing.


Former-commit-id: ca668a3af573009be5ca24cdfe638c3b384d2d7a