SGTools: Fix for removal of symlinked proxies.
It can happen that we have two proxies A and B with the same name but different
CLIDs that nevertheless share the same base class; call the CLID of the
shared base CB.

In such a case, only one of the two proxies will be entered in the store map
for CB, call it A.  If we then try to remove B, the existing code was removing
the CB entry as well, which led to inconsistent proxy refcounts.
Need to check when we're removing symlinks that the entry actually refers
to the correct proxy.

We can also skip the primary CLID in the symlink removal loop, since that's
already been done.



Former-commit-id: 666b028bc8a3bdf70ea39b572cd396ce8dfcc5c2