general fix of config in MuonRecExample (ATLASSIM-3461)
For NSW reco a muonRecFlag called doNSWNewThirdChain needs
to be set to true. In common reco jobs it is false. The NCB
segments are only built during common reco jobs, i.e. if
doNSWNewThirdChain is false. However, the xAOD maker for
these segments is always called. This results in crashes
when doNSWNewThirdChain is set to true. To fix the problem
the xAOD maker needs to be moved inside the config when
doNSWNewThirdChain is false.


Former-commit-id: 45b30e18924c2173c1ab57142d9f7b9fa27ff588