Cleaning up EXOT4 and adding generation option
EXOT4 had a few redundancies in it, which are cleaned up here.

Also adding a 'generations' option to the TruthDecayCollectionMaker.
This allows a given number of generations in the collection, so that
(for example) one can save exactly one generation after a top quark, or
one generation after a tau lepton.


Former-commit-id: 0e108e3d13ee095886b3f932e1e09d89aad0aa07