Modifications to get MCChannelNumber
Trying a few different paths to get the MCChannelNumber from the file.
First up is the old-style event info (should exist in many cases when we
are making an xAOD truth collection, including TRUTH derivation
production).  Second up is the new-style event info (should exist in
some other cases, like reco jobs after digi is fully converted to xAOD
event info).  Third up is the streamer, which was what we used to use
and should be a reasonable fall-back.  Following
EventInfoCnv::massageEventInfo, in the case that the channel number is
0, instead accessing the run number (this is a good thing to do in case
the channel number has not been set).

Former-commit-id: 84874fb43ff0bb6cf856a2ff5231a458ee8899af