SGTools: Embed TransientAddress in DataProxy.
Change TransientAddress to be a member of DataProxy rather than a separate,
owned object.  Allows for better locality of reference and removes
a pointer indirection.  Add move operator and constructor for TransientAddress,
as well as copy/assignment that correctly maintain the reference
counts on m_address.

Also reorder fields of TransientAddress and DataProxy to remove padding
and to bring the most commonly-accessed fields to within the first
cache line.  Remove DataProxy::svcType(), which is not used anywhere.


Former-commit-id: c1021b596917137bf71c6bc71ecc209ec58cbd9d