adjust DCMathSegmentMaker for NSW (ATLASSIM-3462)
For a more detailed debugging of the segment finding using NSW
hits it is required to add more verbose print statements in the
code. Furthermore it is needed to tune the segment finding in
a first rough way to consider hits in the NSW detectors. This is
done by an inflation of the hit errors.


Former-commit-id: 5458f4d1fc4dbd1e54991956f7caea9a34908131