Thinning of InDetForwardTrackParticles container in the primary AOD
https://indico.cern.ch/event/665904/

This MR puts into effect a decision of the OAB, on the recommendation of the (D)AOD size task force, to remove all InDetForwardTrackParticles that are not associated with SiliconAssociatedForwardMuons, from the AOD. See the minutes and slides in the above agenda. The objects are removed via the thinning service, so are still created in the reconstruction.

This modification violates frozen Tier-0.


Former-commit-id: 89879edced6cd032c7d1d33a5c5cc1836449b749