Updating JetCalibTools in preparation for MC16 JMS
The MC16 JMS uses a new JMS parametrization, switching from
(pt vs m vs eta) to (e vs log(m/e) vs eta) or (e vs log(m/et) vs eta)
or (e vs log(m/pt) vs eta).  All of these are now supported.

The code was also updated to prepare for missing LCOrigTopoClusters
in most derivations (as they are transient unless explicitly requested),
which means that jet.getAttributes() is not possible to use.  Instead
of requiring >1 constituent for the JMS, the mass is just required to
be non-zero, which is an equivalent safety check for division by zero.

A new standalone executable was also added, which is useful for
evaluating if the JMS is working as expected.


Former-commit-id: fd4f021bbbbaf42e1bcd613b04ddea98c5c3d867