adjust MuonGeoModel for NSW (ATLASSIM-3467)
For the full details please see the Jira ticket. In short is this
commit and the previous one a cherry-pick of 055324f5ac [formerly ddee1fe55f] and a
modifcation according to commit d896b24321 both in the atlas-muon-
nsw-sim-dev project. They adjust MuonGeoModel to be capable to run
the NSW - in particular the sTGC geometry. The compilation in
release 21.3 was successful.


Former-commit-id: 284056369e145b169b16c801e57fe12cf74308ec