adjust MuonAGDD code for NSW (ATLASSIM-3469)
The two packages MuonAGDDDescription and MuonAGDDBase provide
information regrading geometry and readout structure to other
packages. It was figured out that more/different information
than initially thought needs to become available. The
corresponding values are present in more recent geometry input
files and require an update of the packages to process this data
and make it accessible afterwards. The handlers are very generic
and do not inherit form any Athena class, i.e. the common message
service cannot be used at this point and standard cout is
implemented. For more details please see the Jira ticket. The
software compiles well in release 21.3.


Former-commit-id: d8c8d1df856db43e84784952d1c1b7a85b8a0bcc