EXOT21 trigger strategy update
EXOT21 trigger strategy update to use QCD triggers only on a custom 10% 2015-2016 GRL. This EXOT21 update brings the athena Git repository up to date with the strategy implemented in SVN tag DerivationFrameworkExotics-00-05-50 and following.

GoodRunsLists had to be updated in order for it to be a DerivationFramework skimming tool. The modifications applied here reflect those applied in SVN tag GoodRunsLists-00-01-35 by Will Buttinger.


Former-commit-id: 92ba2b17d15762827bc02ab15ffd1a02fa78cef5