AthenaServices: Fix DP handling in MixingEventSelector::loadAddresses.
MixingEventSelector::loadAddresses was each time creating and registering
a new DataProxy for the MergedEventInfo object.
This worked before, although it was likely a memory leak.
However, with the changes to the organization of DataStore,
addToStore now returns an error if a proxy already exists
for the requested clid/key.  Update MixingEventSelector to create
the proxy only the first time.

Fixes ATN failure in EventLoopTest.


Former-commit-id: fd194fd3d5536bf352110f5df029a21074e005c5