Adding Geometry Configuration Tools
These tools hang off the `DetectorGeometrySvc` and will allow geometry
building from GDML to work properly. The first use-case uses them to
define non-standard materials (replacing the code that was previously
in `DetectorGeometrySvc::BuildExtraMaterials()`) and is in this merge
request. They could also be used to set smartlessness or other tracking
parameters in a neat way (possibly could be an alternate approach for
ATLASSIM-1727). See ATLASSIM-3047.


Former-commit-id: 4cb888e03b3f183717948e0fb10d0912e3b05dde