Fix for ATLASRECTS-4204
The MuonSegmentRegionRecoveryTool constructs a ROI based on the max and min eta and phi positions of the MS hits to determine where to look for missing hits.
To ensure that only MS hits are used to make the ROI, hits with parameters r and z less than 2500 and 4000 respectively are rejected.
In the extremely rare (in that this is the first time I've ever seen it) case that the fit is so terrible that all hits have r<2500 and z<4000, no hits are found.
As a result, the initial values of -1e9 and 1e9 for the max and min are passed to the RoiDescriptor object for both eta and phi, causing a timeout.
(These default values are, of course, selected to be smaller and larger respectively than any possible eta or phi value of the hit.)
The simple fix is to bail if no hits are found: that's what this commit implements.
At some point, we will have to figure out why this fit is so bad, but that will be more complex and this quick fix should allow the data to be recovered in the reprocessing.


Former-commit-id: 82f464efde6e7494bdafed893191d50585551f5c