Forcing IOV check for the first event of the run (ATEAM-439)
IOVSvcTool performs the first round of IOV checking and callback firing in the
"BeginRun". After that it skips this procedure in the "BeginEvent" of the first
event of the run and resumes IOV checking in "BeginEvent" of all subsequent
events. This optimization works fine for serial jobs, but does not work well
in MP where we do "BeginRun" in the master process and all "BeginEvent"-s happen
in workers. As a result no IOV checks are performed for first events of all MP
workers and, consequently, we may miss some IOV callbacks, like it is described
in ATEAM-439.

The MR forces IOV checks for the first event of the run, which results in
checking IOV-s for first events in all MP workers. This should fix the problem
of missing callbacks by introducing a negligible overhead (0.3ms/job) in
serial jobs.


Former-commit-id: 6d269c50fbe52b3bc529ae87d756ad8fb6fa3e2a