LArCalibTools: Compatibility with existing configurations.
LArPedestals2Ntuple and LArRamps2Ntuples were changed to get their inputs
from the conditions store rather than the detector store.
However, this breaks existing configurations, such as used in
the LArCalibProcessing tests, which still put their outputs 
in the detector store.  Updating these configurations to do everything
in the conditions store is not trivial, especially since some of them
rely on in situ modification of objects in the detector store.
Interactions with conditions writing in RegistrationServices also
needs to be considered.

For the time being, modify these two ntuple classes to look for their
inputs in the detector store first, and only after that in the
conditions store.

Fixes ATN failures in LArCalibProcessing.


Former-commit-id: 409606e116a1d7a5cee4b9c4a697a43bc41c874f