Adding several fields to TruthMetaData for ATLASG-1324
In ATLASG-1324 we pointed out that the ROOT-readable metadata that comes
with all the derivations is missing several key items in order to make
it maximally useful for analyzers, like the information about what
generator is being used, what tune, what LHE generator, and so on.  This
merge requests adds that information in.  If the information is not
available, then we get back a blank.  Of course, this is only as good as
the event generation transform's ability to set consistent metadata, but
that should be pretty good in MC16 already.

This allows a number of nice things, like the automatic configuration of
b-tagging MC/MC scale factors.  We should really try to make it
available in time for the large-scale derivation reprocessing coming up!


Former-commit-id: 1bbdf87899d85d2f7b5aa990e4465cce1aedefde