Fixing two minor bugs in TRUTH common setup
First bug: if TruthWZ jets were in the input file, then TruthDressedWZ
jets would not be built.  Of course, no one is using this collection
yet, and no one is really running into this situation, but ok, better to
fix it now.

Second bug: in xAOD making, the truth record is thinned; in particular
some particles from the Geant4 record are removed.  These can create
null pointers in the truth record in decay products.


Former-commit-id: 0cd6c37e6f89662a541759db98f4206fe75a5cbd