remove deprecation warnings
It is not that the warnings are wrong per se: These classes are
inherently thread-hostile and code that relies on them needs to be
changed.  Any code that relies on TResult will likely crash and burn
for nthreads > 1, and any code that relies on TAccept will likely
experience data corruption for nthreads > 1.

The reason for disabling the deprecation warnings is that we get too
many users, and we essentially get swamped by deprecation warnings.
As such I commented out the deprecation warnings (but left them in
with an explanation), as we will have to remove these classes before
we can run in multi-threaded mode.

See discussion here: https://gitlab.cern.ch/atlas/athena/merge_requests/4801


Former-commit-id: a7ca3494c1ce7f054b4550685ccaeb55539f5b88