Forcing IOV check for the first event of the run (ATEAM-439)
IOVSvcTool performs the first round of IOV checking and callback firing in the "BeginRun".
After that it skips this procedure in the "BeginEvent" of the first event of the run and
resumes IOV checking in "BeginEvent" of all subsequent events. This optimization works
fine for serial jobs, but does not work well in MP where we do "BeginRun" in the master
process and all "BeginEvent"-s happen in workers. As a result no IOV checks are performed
for first events of all MP workers and, consequently, we may miss some IOV callbacks,
like it is described in ATEAM-439.

This MR forces IOV checks for the first event of the run only in MP jobs. In order to do that,
IOVSvcTool relies on the value stored in Gaudi::ConcurrencyFlags::numProcs. AthenaMP was
modified such that now it is setting the value of that Gaudi flag at initialize(), and, finally,
Gaudi was modified in order to allow AthenaMP make changes in ConcurrencyFlags
(see https://gitlab.cern.ch/atlas/Gaudi/merge_requests/154)


Former-commit-id: 7d13c888a241c65c85910f0e93cee31ec4e73686