change it_GridTools to stand-alone executable
Since it_GridTools doesn't work in the CI system (missing grid
certificate) I switched this to a stand-alone executable for now.
This should be switched back, once we have a way to run this either in
CI or in the nightlies.

Made a ticket for it here: https://its.cern.ch/jira/browse/ATLINFR-2063


Former-commit-id: 8b0c5928b66df3a377c52c45a0e874b3fd060f79