Accounting for photon radiation
Some of these vertices have multiple leptons as well as a few outgoing
photons (assuming these are FSR).  Have to take these into account when
deciding whether the vertex is a VVV vertex.  Hard to imagine that
there's something more sinister going on when we have six leptons coming
out of a single vertex.


Former-commit-id: 3d3999e84565ac129becd479cf2d86a9ba521c4e