Adding Proposed MC18 Truth Strategy Configurations
The `MC18` and `MC18LLP` truth strategies are the same as the
`MC15aPlus` and `MC15aPlusLLP` truth strategies used in the MC15
and MC16 campaigns except that. The first secondary barcode is
`1000001` rather than `200001` and that the barcode is incremented
by `10000000` rather than `1000000` for each generation (each time
a particle interacts and survives).

This is related to the discussion in ATLASSIM-3103.


Former-commit-id: 45b5b981aafd01c7f32c9c08f3bb8318be1109a1