Drop usage of SecondaryTracksHelper (ATLASSIM-3492)
The `SecondaryTracksHelper` is not thread-safe, so all usage of
the class has been removed and replaced by direct calls to
`G4Step` methods. These changes mirror those made for the old
FADS-based truth strategies for ATLASSIM-2409.
The `SecondaryTracksHelper` class has been removed from the release
as it no longer has any clients.
This change should fix ATLASSIM-3492 once it is swept to `master`.
It makes sense to add it to `21.0` first to keep the `Simulation`
code as consistent as possible between branches.


Former-commit-id: 0529cca2eba726d6448acdd0759552d69f436bd2