Added api module to provide lists of unprescaled triggers
The module reads the TriggerDB to provide lists of unprescaled triggers
per type and/or period.
It can provide also information for what is expected to be unprescaled
in the future.
When TriggerMenu is built it produces lists for the current unprescaled
triggers, which can be used by other clients


Former-commit-id: 009d306a5359b3430b758b116f0863ecdf1cab7f