Adding support for 3D histograms (ATLJETMET-700)
So far, JetCalibTools has used 2D histograms for separate eta bins.
This is detrimental to performance, as it means that there is no
interpolation across eta.  This change adds support for 3D histograms
in multiple pieces.  This is demonstrated by a full implementation
for the JMS calibration.  A similar implementation for the GSC
is left for the future.

RootHelpers is a file which provides support for proper interpolation
in 3D with a TH3, and is needed even though it is a near copy of the
root TH3 interpolation code, for two key reasons:

1. Interpolate makes use of FindBin, and FindBin is *not* a const
function in ROOT.  This is because FindBin is allowed to modify the
bin range *if* a given flag is set.  We never want this to happen,
so I've removed that functionality, and made a version which is
const with respect to the histogram it is acting upon.

2. TH3 interpolate does not work for edge bins.  By this I mean the
first/last real bin, not the underflow/overflow bins.  In such cases,
the TH3 should fall back on 2D interpolation for the dimensions which
are not on the edge.  This is implemented in this commit.

The above RootHelpers are a direct copy/paste from where I first
implemented them for the JetUncertainties package, in the
UncertaintyHistogram class.


Former-commit-id: b909422a1dac4c2e36f7be59015bcef6c9073a76