Added support for multiple alignments in flight to GeoModelKernel
GeoModelKernel was modified in order to support multiple sets of alignment
constants in flight in AthenaMT jobs. The "dynamic" part of GeoModel tree
(alignment deltas and full physical volume position caches) was decoupled
from the "static" part. The dynamic data will be stored into a new entity:
GeoVAlignmentStore. The interface class was added to GeoModelKernel, while
the concrete implementation was added to GeoModelUtilities package.
The latter one is expected to be used by the alignment clients.

The changes have no effect on the client code. No immediate actions
are required from the clients. They can start the migration to the new
mechanism adiabatically.

The changes have been tested by doing a full rebuild of the master branch
and running several q-tests afterwards.


Former-commit-id: dc76fa91846bc052764255a98763133a88c390c8