Further fixes for ATLASRECTS-3796
The original fix for missing truthParticleLinks implicitly assumed that the xAOD::TrackParticle always had an attached truthParticleLink.
In turn, the code for adding the truthParticleLink to the xAOD::TrackParticle implicitly assumed that the trackParticle always had an attached trackLink().
However, recent studies on 21.0-mc16d by Jovan seem to indicate that at least one of these assumptions is violated.
This fix should hopefully ensure that a dummy truthParticleLink is always added to the trackParticle and the Muon object no matter what.


Former-commit-id: 45a1b270a93eddf27068b5edb1fbe29ee470049b