Adding inter-collection navigation for truth particles
One long-standing request is for the ability to navigate between the
small truth collections that are provided in TRUTH3 and similar
derivation outputs.  This merge requests makes the navigation possible
by adding links between the collections, for both the children and the
parents of all the particles.  The navigation is done with links
(previously there was a saving of the PDG ID of the mother, but that's
pretty insufficient for some cases) and the code can handle particles
that appear in multiple collections (it decorates them all).

This comes with some nice side-benefits.  First, if you want to hop
around the collections, that's no problem.  Second, as the particles are
ordered by priority, you are free to include the full truth collection
(if it exists) in the list of collections to use, and in that case there
will be links assigned to all known parents and children.

Tested in a nightly that a reasonable number of links are coming out
(e.g. usually one parent, though the code can handle multiple parents,
and no children for neutrinos, etc).


Former-commit-id: ebee3c26d12600967b80117837e007b840ce7ee1