Update of the SUSY11 DerivationFramework config (ATLSUSYDPD-1081)
In order for the SUSY11 derivation to be more future proof the
trigger skimming is changed to one using regular expressions instead
of a hard coded list.
The navigation thinning uses the same trigger expression.
BTaggingContainers are removed from the output as they have been
deemed unnecessary for the purposes of this derivation.


Former-commit-id: ed3f80f5a2d18230d39511332eca4d7b540b7d42