Switch IEnvelopeDefSvc to use DeclareInterfaceID
This change is a further patch for the changes in !4814.

`AFIIEnvelopeDefSvc` uses the `extends<...>` pattern to declare its
interface, but the interface class `IEnvelopeDefSvc` still used the
old incompatible (at runtime) approach tp declare its interface. This
cause `ATLFASTII` jobs to fail during initialization. Updating
`IEnvelopeDefSvc` to use `DeclareInterfaceID` fixed this issue.
Then `DetDescrDBEnvelopeSvc` had to be updated to use `extends<...>`
to make it compile.

This is part of the changes for ATLASSIM-3439.


Former-commit-id: d3301bb71b38c0a301c214cfa90193a1b51880d5