LArRawConditions: Add a non-const iterator to LArConditionsContainer.
Also requires adding a non-const iterator to LArConditionsContainerDB.
The original comments said that a non-const iterator wasn't desired.
However, it is already possible to modify the container contents via
set() and getNonConst(), so this isn't really opening up any new holes.
And it allows getting rid of some const_casts that were put in downstream code.

Also added a unit test for LArConditionsContainerDB.


Former-commit-id: 94db7ada221e35e8f464e871ab1a0f48aa5b59ab