StoreGate: Avoid redundant retrieves of store service.
VarHandleKey::initialize() was calling retrieve() on the handle to the referenced
store.  However, retrieve() unconditionally queries the service manager
to find the referenced service.  In the case where a var handle is initialized
from a key, the service handle to the store is already valid, so this
is redundant.  Change to using isValid() instead, which queries the service
manager only if the pointer in the service handle is not already valid.

Former-commit-id: ff8fa6992a647bb8027e0041a1b3c58e81bfd13f