Further fix for ATLASRECTS-3796
Found another way in which a muon could fail to have a truthLink applied.  The code only adds links to status=1 truthParticles, so if the trackParticle is linked to a truthParticle with status!=1, the code will not find it.  However, since the truthLink will be valid, no dummy information was being written.  This should now be fixed, and dummy information written in the truthLink is valid but the truthParticle has bad status.


Former-commit-id: abd42184ef54d15e5d872fc7f9c31e9d4759cb1c