Fixing bug in one mass strategy (ATLJETMET-700)
In the process of further validation, a bug was found in one of the
supported mass strategies (a key break was missing from a switch
statement).  This commit fixes that problem.

This MR also adds a new validation script, but that is coming
along for the ride.  The urgent part is the fix mentioned above.


Former-commit-id: af5e79bef5ed58023a87f23aa9cb4f95073047bd