Adding truth born lepton collection
Last long-standing request, for a special born lepton collection in the
truth derivations.  Code based on H->ZZ->4l born selection code, and
handles some generator special cases already.

One outstanding item: the metadata issue.  I'd welcome comments on this
one.  In a nutshell, I want to get the generator information in as
efficient a way as possible.  The TagInfo metadata seems not to be
available in the detStore when this tool is initialized.  The
xAOD::TruthMetaData should be there (as it is created when the
xAOD::TruthParticleCollection is made, on which this code all relies),
but it's not in storegate as far as I can tell.  I'm honestly not sure
why that would be.

Other than that, the code seems to be behaving.


Former-commit-id: 7d9986ca4389e3b660f7d906aa0d2f071cd29e44