Moving to functions for tau truth kernel
This change has two related effects.  What was in TauTruthCommon is now
wrapped up in a function.  All clients have been migrated to use the
function form, and several clients were identified that don't actually
need to be calling this function at all (they get it for free when using
the common tools).

The first good benefit is we avoid the `import *` syntax, and treat the
module as a real module that has a function we can call rather than as a
replacement for `include`.

The second good benefit is that we can optionally pass this function an
algSequence in which to schedule the truth tools.  At the moment, the
truth code is the only client of this functionality, but it allows the
scheduling of private instances of the kernels that could run (for
example) after skimming choices have been made, rather than public
kernels that are always a part of the main sequence.  This satisfies a
request of the SUSY group, to ensure that the recommended workflow
(private kernels) really is possible, even if it's not heavily used.

Tested a TRUTH and EXO derivation to make sure that things are behaving
after these changes, and everything seems to be behaving normally.  This
change was not discussed at any length with the tau group, though, so it
would be useful to have a comment from someone on that end before
merging this request in.


Former-commit-id: f3e5cae04c2ab2a533d4a25edcc0e3a91d14b0dd