New LLP Truth Derivation
Updated truth decay collection maker to spit out a collection of BSM
particles and their decays.  This can be used (for example) for
long-lived SUSY particles, dark photons, and so on (note that if the
SUSY particles are strongly produced, this collection could get large).
In the case of LLP analyses, the fact that vertices are also included
here means that it's actually possible to look at the decay positions of
the BSM particles, which is important for a number of truth studies.

Making TRUTH4 into TRUTH3 + this new LLP truth particle collection.
TRUTH4 previous client was H->ZZ, who we're working with on a new
derivation (after a couple of merge requests are added, it should be
straightforward); they are currently working from TRUTH1.

This will form a basis for future discussions of truth derivations for
LLP analyses.


Former-commit-id: eaefdcfe3c0a5c687cad05bbe10ab070b5eae9d4