SGComps: Output of a WriteHandle should supercede something from the input file.
If we have a WriteHandle for something that is present in the input file,
then that shouldn't be an error; instead, the data from the input file
should simply not be visible.  This is implemented by the AddressRemappingSvc.
It goes over the list of all WriteHandles and removes any matching data items
from the TAD lists.


Former-commit-id: 5c902e18f9ce2c4a3be414b9a0f694b94b50de2a