svnpull of updates from `ISF_Geant4Event-00-00-07` and `ISF_Geant4Tools-01-00-00`
Some of the updates from these tags were already in `master`, but the udpates
to `ISFG4GeoHelper` and `PhysicsValidationUserAction` were missing.
The `PhysicsValidationUserAction` updates just tidies up the `MessageSvc` usage.
The `ISFG4GeoHelper` update is part of the changes for ATLASSSIM-3000
to make the ISF truth service work in G4Atlas with and without the
cavern volume. It also contains some readability improvements.


Former-commit-id: 00cabbf90040dd38aeb72881b41691462188cd5b