Tidying of `TruthStrategyManager` and `AtlasTrajectory` classes
The main purpose of this commit is to remove the
`TruthStrategyManager::InitializeWorldVolume()` method, which
was not thread-safe. It has been replaced by a configurable
property of the `AthenaTrackingActionTool` class, which is then
passed to the `AthenaTrackingAction` and then to each
`AtlasTrajectory` object which is created.
At the same time the redundant `m_numCurrentSec` member variable
of `AtlasTrajectory` has been removed. The remaining
`TruthStrategyManager::CreateTruthIncident` method has been made
`const`. In the future it may be possible to merge this code into
`AtlasTrajectory::AppendStep` by following the example of the
`ISFTrajectory` code. These changes are part of the fixes
for ATLASSIM-3492.


Former-commit-id: ec5b7938c3c51d30fae779a3d14e23e9c70ed1ec