GeneratorFilters: const fixes.
DataVector has had a long-standing bug in which it has been possible
to get a non-const pointer from a const_iterator.
Fix instances where this package was relying on that to compile.

The class LeadingDiBjetFilter is further problematic as it was taking
advantage of this to modify the weights in an existing GenEvent object
in StoreGate.  However, there's nothing currently using this class,
so just remove it.


Former-commit-id: a74747a7dde32cdac3888c30eba970612a7afe2b