Make Calibration Area tag a property of the tool
The default one is the latest tag (00-04-81) containing the final
MCJES+GSC configs for R=0.4 EMTopo, LCTopo and p-flow jets.
Once new tags become available the default option will be updated.
This allow to release new recommendations faster, users will simply need
to update the tag when initializing the tool. Except when changes in the
code are needed.


Former-commit-id: 5c42033661f047cacc1697dcd84dec985b1600a1