Generalise chain patterns somewhat
Due to trivial changes in the naming of certain chains, the patterns that we use will not
match some of these changes, eg if we have a HLT_mb_idperf chain, then our usual pattern of
HLT_mb_.*_idperf where we normally expect something between the mb and idperf, eg HLT_mb_sp200_idperf
for instance, will not work because of only the single "_".

This change replaces ".*_idperf" by ".*idperf" so that these changes in the naming convention
will be automatically handled.


Former-commit-id: 21e1d4d35c87470259869d892e228d9130e48cfa